add structure, localize enemies with is_pacman flag
