+++ pacman-contest/src/contest/layouts/myTestCapture.lay
@@ -1,0 +1,7 @@
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%.1   .      .    .            2             %
+%.%%.%.%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%.%.%%.%
+%......     ... ...                   4      %
+% %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% %
+%        3%%                                 %
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
